Moved tests files under src/tests folder
